pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "market-data-mcp"
version = "0.1.0"
description = "Crypto market data MCP server (CoinGecko + Fear&Greed + CryptoPanic)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.23.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.21"]

[project.scripts]
market-data-mcp = "market_data_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/market_data_mcp"]
